Create match_summary.lua
- fixed the file name (it is a MatchSummary Module and not a MatchGroup Module))
- copied from the live version on the wiki
